% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landscapes.R
\name{extract_landscapes}
\alias{extract_landscapes}
\title{Extract multi-scale landscapes}
\usage{
extract_landscapes(r, p, buffers, strategy = "nested", ...)
}
\arguments{
\item{r}{A \verb{Raster*} or \code{SpatRaster} object.}

\item{p}{A set of sampling points (\code{sf}, \code{SpatVector}, or \verb{SpatialPoints*}).}

\item{buffers}{A numeric vector of buffer sizes (in map units).}

\item{strategy}{The extraction strategy: \code{"nested"}, \code{"decouple"} or \code{"partial_decouple"}.}

\item{...}{Reserved for future use.}
}
\value{
An object of class \code{lsma}.
}
\description{
Cut a raster around sampling points using different buffer radii and strategies. Returns a structured object of class \code{lsma}.
}
