% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodel.R
\name{multimodel}
\alias{multimodel}
\title{Multimodel}
\usage{
multimodel(
  landscape_metrics,
  bio_data,
  model_type,
  model_formula = "responses ~ predictors",
  response_vars = NULL,
  model_args = NULL,
  quiet = FALSE
)
}
\description{
TODO: Describe this function.
}
