% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_buffers.R
\name{nested_buffers}
\alias{nested_buffers}
\title{Nested buffer extraction}
\usage{
nested_buffers(r, p, buffers)
}
\arguments{
\item{r}{A SpatRaster object.}

\item{p}{A SpatVector with sampling points.}

\item{buffers}{Numeric vector of buffer radii.}
}
\value{
A named list of SpatRaster objects.
}
\description{
Extracts nested buffers from a raster based on multiple radii.
}
\keyword{internal}
