% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_buffers.R
\name{partial_decouple_buffers}
\alias{partial_decouple_buffers}
\title{Partially decoupled buffers}
\usage{
partial_decouple_buffers(r, p, buffers, keep_buffer)
}
\arguments{
\item{r}{A \code{SpatRaster}.}

\item{p}{A \code{SpatVector} of points.}

\item{buffers}{A numeric vector of buffer distances (sorted ascending).}

\item{keep_buffer}{The buffer to isolate (must be one of the values in \code{buffers}).}
}
\value{
A list with \code{rasters} and \code{buffers}.
}
\description{
Create buffers where one focal buffer is isolated, and all others are grouped.
}
