% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.lsma}
\alias{summary.lsma}
\title{Summary for lsma objects}
\usage{
\method{summary}{lsma}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lsma}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Display a summary of an object of class \code{lsma}, including metadata and structural information.
}
