% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_buffers.R
\name{partial_decouple_buffers}
\alias{partial_decouple_buffers}
\title{Partial decoupled buffer extraction}
\usage{
partial_decouple_buffers(r, p, buffers)
}
\arguments{
\item{r}{A SpatRaster object.}

\item{p}{A SpatVector with sampling points.}

\item{buffers}{Numeric vector of buffer radii.}
}
\value{
A named list of SpatRaster objects.
}
\description{
Extracts partially decoupled buffers based on a subset of radii.
}
\keyword{internal}
